Push to main

+++ controllers/eventController.js
@@ -1,4 +1,20 @@
 const Event = require('../models/Event');
+const Booking = require('../models/Booking');
+
+const escapeRegex = (value) => {
+  return String(value).replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+};
+
+const EVENT_UPDATE_FIELDS = [
+  'title',
+  'description',
+  'category',
+  'venue',
+  'date',
+  'time',
+  'seatCapacity',
+  'price'
+];
 
 const getEvents = async (req, res, next) => {
   try {
@@ -6,7 +22,7 @@ const getEvents = async (req, res, next) => {
     const filter = {};
 
     if (category) {
-      filter.category = new RegExp(`^${category}$`, 'i');
+      filter.category = new RegExp(`^${escapeRegex(category)}$`, 'i');
     }
 
     if (date) {
@@ -70,7 +86,11 @@ const updateEvent = async (req, res, next) => {
       throw new Error('Seat capacity cannot be lower than already booked seats');
     }
 
-    Object.assign(event, req.body);
+    EVENT_UPDATE_FIELDS.forEach((field) => {
+      if (req.body[field] !== undefined) {
+        event[field] = req.body[field];
+      }
+    });
     const updatedEvent = await event.save();
 
     res.json(updatedEvent);
@@ -86,6 +106,13 @@ const deleteEvent = async (req, res, next) => {
     if (!event) {
       res.status(404);
       throw new Error('Event not found');
+    }
+
+    const bookingExists = await Booking.exists({ event: event._id });
+
+    if (bookingExists) {
+      res.status(400);
+      throw new Error('Cannot delete an event with existing bookings');
     }
 
     await event.deleteOne();
